Rename fastapi stub and update tests
