+++ .dockerignore
@@ -14,3 +14,4 @@ tmp/
 tools/poppler/Library/include/
 tools/poppler/Library/lib/
 tools/poppler/Library/share/
+fastapi_stub/
