+++ README.md
@@ -116,6 +116,8 @@ bash scripts/download_examples.sh
 pytest -q
 ```
 
+> В каталоге имеется упрощённый пакет `fastapi_stub`, который используется **только** для запуска тестов. Для реального запуска приложения требуется установить полноценный FastAPI из `requirements.txt`.
+
 ---
 
 ## Лицензия
