+++ fastapi_stub/staticfiles.py
