+++ fastapi_stub/templating.py
